[feature] Listmaster notifications can now be grouped.
If, for a specific robot, several notifications are issued within a short period
of time with the same operation code only the first ones are sent, the next ones
are stacked. The stacking ends when no notifications are issued anymore, or when
the first one is too old. All the stacked notifications are then sent as attachments
of a single message.

Internal settings :
  - Stack if more than 3 notifications with the same operation code are issued
    for the same robot
  - Send stacked notifications if no new one satisfying the condition above was
    received for more than 30 seconds
  - Send stacked notifications if the oldest one was issued more than 60 seconds ago