[-bug] Message distribution fails when tracking feature is enabled with Postfix MTA.  Because Postfix clone of sendmail command doesn't allow "-V envid" but "-Venvid".
Note:
Postfix 2.3 or later is required to use tracking feature.
Recent versions of Sendmail and Courier also supports it.
Exim and qmail seem not to be supported.

Additionally, options given to sendmail command will be terminated by "--" (double hyphen) to handle recipient addresses begining with "-".  This feature is available on Sendmail V8 or later and clones of it (at least Courier, exim, Postfix and qmail).